fix bug

+++ crm/install.py
@@ -160,8 +160,9 @@ def add_default_salutation():
 	if not is_edited:
 		try:
 			for salutation_sys in salutations_sys:
-				salutation_doc = frappe.get_doc('Salutation', salutation_sys)
-				salutation_doc.delete()
+				if frappe.db.exists('Salutation', salutation_sys):
+					salutation_doc = frappe.get_doc('Salutation', salutation_sys)
+					salutation_doc.delete()
 			for salutation_sys in ["Mr", "Madam", "Prof", "Miss"]:
 				if frappe.db.exists("Salutation", salutation_sys):
 					continue
@@ -185,8 +186,9 @@ def add_default_gender():
 	if not is_edited:
 		try:
 			for gender_sys in genders_sys:
-				gender_doc = frappe.get_doc('Gender', gender_sys)
-				gender_doc.delete()
+				if frappe.db.exists('Gender', gender_sys):
+					gender_doc = frappe.get_doc('Gender', gender_sys)
+					gender_doc.delete()
 			for gender_sys in ["Female", "Male", "Other"]:
 				if frappe.db.exists("Gender", gender_sys):
 					continue
